Update backup.tf
This patch adds the destination vault name as an output from the `aws-backup-destination` module.  This is needed because you can't pass an ARN to the [`aws_backup_vault` terraform data source](https://registry.terraform.io/providers/hashicorp/aws/latest/docs/data-sources/backup_vault), so if you want to refer to the vault from a different config you're a bit stuck.